# (internal): Simplify `org-roam--list-files` implementation.