Use destructive elisp functions to optimize for speed and memory
usage.

Refs #104

###### Motivation for this change

slight speedup:
Original:
```
ELISP> (benchmark 1000 '(org-roam--list-files org-roam-directory))
"Elapsed time: 29.403994s (6.193575s in 122 GCs)"
```
Commit:
```
ELISP> (benchmark 1000 '(org-roam--list-files org-roam-directory))
"Elapsed time: 27.186161s (5.973992s in 117 GCs)"
```