@juergenhoetzel 2020-04-19
BTW removing `file-truename` for each file would result in a huge speedup:

```
ELISP> (benchmark 1000 '(org-roam--list-files org-roam-directory))
"Elapsed time: 3.529275s (0.933762s in 18 GCs)"
```
Only downside: Symbolic linked org files don't get resolved.