@jethrokuan 2020-04-19
The elisp implementation of finding these files should be a backup, this is tracked in an old issue #104