@juergenhoetzel 2020-04-20
> The regexp for this needs to be built from org-roam-file-extensions.

Thanks: I amended the commit.

> I'm overall in favour of this, but the old code had a dir-ignore-regexp that tries to be smart by ignoring certain hidden directories (like those that start with ., e.g. .git), not sure if the performance is slow when there are such directories in org-roam-directory with a lot of objects.

I was able to remove the `file-truename` bottleneck (from `directory-files-recursively` docstring):


> Each file name appears in the returned list in its absolute form.


This results in a huge performance speedup:

Original Code:

```
ELISP> (benchmark 1000 '(org-roam--list-files org-roam-directory))
"Elapsed time: 21.594839s (7.690849s in 156 GCs)"
```
Commit:

```
ELISP> (benchmark 1000 '(org-roam--list-files org-roam-directory))
"Elapsed time: 6.259151s (1.435158s in 29 GCs)"
```
I would guess the ".git" special handling would not improve performance, because git uses pack-files with corresponding indices for large repos and the elisp code would be more complex.