@timorunge 2020-05-06
@juergenhoetzel, @jethrokuan: If you're using tools like e.g. [Syncthing](https://syncthing.net) which has local versioning (the data is stored in a hidden folder) you will be able to see all versioned files. 

The correct RegEx to exclude hidden folders would be something like `^[^.](.*).(?:(org))(?:.gpg)?` but due to my limited lisp knowledge I was not able to get it working...